llm-judge-interview-2/llm-judge-interview-2/requirements.txt
jsonargparse
pymongo
pandas
voyageai
scikit-learn
dspy-ai
jinja
